🧹🧹🧹
Some very much needed refactoring
