+++ .github/workflows/clean-sort.yml
@@ -1,44 +1,11 @@
-name: Main
+name: Clean & sort
 
 on:
   push:
     branches: [ master ]
 
 jobs:
-  make-template:
-    name: Make template
-
-    runs-on: ubuntu-latest
-
-    permissions: write-all
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-
-      - name: Setup java
-        uses: actions/setup-java@v3
-        with:
-          distribution: oracle
-          java-version: 17
-
-      - name: Compile and run java
-        run: |
-          chmod +x "./.github/compile_run.sh"
-          "./.github/compile_run.sh" template
-
-      - name: Commit & push
-        run: |
-          git config --local user.name actions-user
-          git config --local user.email "actions@github.com"
-          if ! git diff --exit-code; then
-            git add -A
-            git commit -m "Make template action"
-            git push -f origin master
-          fi
-
   clean-sort:
-    needs: make-template
     
     name: Clean and sort files
