+++ cleanFiles/src/JsonObject.java
@@ -1,0 +1,6 @@
+interface JsonObject {
+
+    public boolean isEmpty();
+
+    public JsonObject removeEmptyFields();
+}
